Move functions a bit
