+++ examples/react-three-xr-controllers/package.json
@@ -1,6 +1,7 @@
 {
     "dependencies": {
-        "@react-three/xr": "workspace:^"
+        "@react-three/xr": "workspace:^",
+        "@pmndrs/xr": "workspace:^"
     },
     "scripts": {
         "dev": "vite --host"
